Suggestion on improving code
Script contains three loops for importing, unzipping and connecting to data base which can be conducted into single loop but it will reduce the editable simplicity which works in block process flow.

code snippet can be added to make sure that unavailability of file is because of holiday. this meta data will be useful must be. timestamp flow will act as index here.


Problems:
Data need to be cleaned,
anomaly of adjusted price need to checked